fix: personal account settings form handling
