+++ src/app/cloud/settings/client_page.tsx
@@ -10,69 +10,22 @@ import FormProcessing from '@forms/FormProcessing'
 import FormSubmissionError from '@forms/FormSubmissionError'
 import Submit from '@forms/Submit'
 import { OnSubmit } from '@forms/types'
-import { useRouter } from 'next/navigation'
 
 import { Button } from '@components/Button'
 import { Gutter } from '@components/Gutter'
 import { Heading } from '@components/Heading'
-import { Message } from '@components/Message'
 import { ModalWindow } from '@components/ModalWindow'
 import { useAuth } from '@root/providers/Auth'
+import { DeletionConfirmationForm } from './DeletionConfirmationForm'
 
 import classes from './page.module.scss'
 
 const modalSlug = 'delete-account'
 
-const ModalContent = () => {
-  const { closeModal } = useModal()
-  const [hasEmail, setHasEmail] = React.useState(false)
-  const [hasPW, setHasPW] = React.useState(false)
-
-  return (
-    <>
-      <Heading marginTop={false} as="h5">
-        Are you sure you want to delete your account?
-      </Heading>
-      <Message className={classes.warning} error={'Deleting your account cannot be undone.'} />
-      <p>
-        Team ownership will be transferred to another team member where possible. If no other team
-        members exist, the team and associated projects / deployments will be
-        <strong> permanently deleted</strong>.
-      </p>
-      <p>To proceed re-enter your account details below:</p>
-      <Text
-        className={classes.emailInput}
-        label="Email"
-        path="modalEmail"
-        required
-        onChange={value => {
-          setHasEmail(Boolean(value))
-        }}
-      />
-      <Text
-        label="Password"
-        path="modalPassword"
-        type="password"
-        required
-        onChange={value => {
-          setHasPW(Boolean(value))
-        }}
-      />
-      <div className={classes.modalActions}>
-        <Button label="cancel" appearance="secondary" onClick={() => closeModal(modalSlug)} />
-        <Submit label="delete my account" appearance="danger" disabled={!(hasEmail && hasPW)} />
-      </div>
-    </>
-  )
-}
-
 export const SettingsPage = () => {
-  const { user, updateUser, login } = useAuth()
+  const { user, updateUser } = useAuth()
   const { openModal } = useModal()
-  const [loading, setLoading] = React.useState<boolean>(false)
   const [formToShow, setFormToShow] = React.useState<'account' | 'password'>('account')
-  const [success, setSuccess] = React.useState<string | null>(null)
-  const router = useRouter()
 
   const handleSubmit: OnSubmit = useCallback(
     async ({ data, dispatchFields }): Promise<void> => {
@@ -99,65 +52,25 @@ export const SettingsPage = () => {
           ],
         })
 
-        return
+        throw new Error('Please confirm that your passwords match and try again')
       }
 
       try {
         await updateUser({
-          name: data.name,
-          email: data.email,
-          password: data.password,
+          name: data?.name,
+          email: data?.email,
+          password: data?.password,
         })
 
-        setSuccess('Your account has been updated')
+        toast.success('Your account has been updated')
+        setFormToShow('account')
       } catch (err) {
         const message = err?.message || `An error occurred while attempting to update your account`
         console.error(message) // eslint-disable-line no-console
-        setSuccess(null)
         throw new Error(message)
       }
     },
     [updateUser],
-  )
-
-  const deleteAccount = React.useCallback(
-    async ({ data }) => {
-      if (user) {
-        if (data.modalEmail !== user.email) {
-          toast.error('Email provided does not match your account, please try again.')
-          return undefined
-        }
-
-        try {
-          const confirmedUser = await login({
-            email: data.modalEmail as string,
-            password: data.modalPassword as string,
-          })
-
-          if (confirmedUser && confirmedUser.id === user.id) {
-            try {
-              const req = await fetch(
-                `${process.env.NEXT_PUBLIC_CLOUD_CMS_URL}/api/users/${user.id}`,
-                {
-                  method: 'DELETE',
-                  credentials: 'include',
-                },
-              )
-
-              if (req.status === 200) {
-                toast.success('Your account was successfully deleted.')
-                router.push('/logout')
-              }
-            } catch (e) {
-              toast.error('There was an issue deleting your account. Please try again.')
-            }
-          }
-        } catch (e) {
-          toast.error('Incorrect email or password.')
-        }
-      }
-    },
-    [login, router, user],
   )
 
   return (
@@ -167,7 +80,7 @@ export const SettingsPage = () => {
       </Heading>
       <Grid>
         <Cell cols={6} colsM={8}>
-          <p>
+          <p className={classes.description}>
             {formToShow === 'account' && (
               <Fragment>
                 {'To change your password, '}
@@ -185,7 +98,7 @@ export const SettingsPage = () => {
             )}
             {formToShow === 'password' && (
               <Fragment>
-                {'Change your password below. '}
+                {'Change your password below, or '}
                 <button
                   className={classes.viewButton}
                   type="button"
@@ -193,46 +106,19 @@ export const SettingsPage = () => {
                     setFormToShow('account')
                   }}
                 >
-                  Cancel
+                  cancel
                 </button>
                 {'.'}
               </Fragment>
             )}
           </p>
-          <Form
-            className={classes.form}
-            initialState={{
-              name: {
-                value: user?.name,
-                initialValue: user?.name,
-                errorMessage: 'Please enter a name',
-              },
-              email: {
-                value: user?.email,
-                valid: Boolean(user?.email),
-                initialValue: user?.email,
-                errorMessage: 'Please enter a valid email address',
-              },
-              password: {
-                value: '',
-                initialValue: undefined,
-                errorMessage: 'Please enter a password',
-              },
-              passwordConfirm: {
-                value: '',
-                initialValue: undefined,
-                errorMessage: 'Please confirm your password',
-              },
-            }}
-            onSubmit={handleSubmit}
-          >
-            <Message success={success} />
+          <Form className={classes.form} onSubmit={handleSubmit}>
             <FormSubmissionError />
             <FormProcessing message="Updating profile, one moment" />
             {formToShow === 'account' && (
               <>
-                <Text path="name" label="Your Full Name" />
-                <Text path="email" label="Email" required />
+                <Text path="name" label="Your Full Name" initialValue={user?.name} />
+                <Text path="email" label="Email" required initialValue={user?.email} />
                 <Text
                   value={user?.id}
                   label="User ID"
@@ -243,8 +129,14 @@ export const SettingsPage = () => {
             )}
             {formToShow === 'password' && (
               <>
-                <Text type="password" path="password" label="Password" />
-                <Text type="password" path="passwordConfirm" label="Password Confirm" />
+                <Text type="password" path="password" label="Password" required initialValue="" />
+                <Text
+                  type="password"
+                  path="passwordConfirm"
+                  label="Password Confirm"
+                  required
+                  initialValue=""
+                />
               </>
             )}
             <div className={classes.buttonWrap}>
@@ -266,10 +158,12 @@ export const SettingsPage = () => {
         </Cell>
       </Grid>
       <hr className={classes.hr} />
-      <Heading element="h2" as="h6">
+      <Heading element="h2" as="h6" marginTop={false} marginBottom={false}>
         Delete account
       </Heading>
-      <p>Deleting your account is permanent and cannot be undone.</p>
+      <p className={classes.description}>
+        Deleting your account is permanent and cannot be undone.
+      </p>
       <Button
         className={classes.deleteAccount}
         label="Delete account"
@@ -279,9 +173,7 @@ export const SettingsPage = () => {
         }}
       />
       <ModalWindow className={classes.modal} slug={modalSlug}>
-        <Form onSubmit={deleteAccount}>
-          <ModalContent />
-        </Form>
+        <DeletionConfirmationForm modalSlug={modalSlug} />
       </ModalWindow>
     </Gutter>
   )
