+++ src/app/cloud/settings/page.module.scss
@@ -1,8 +1,18 @@
 @use '@scss/common.scss' as *;
 
+.buttonWrap {
+    display: flex;
+    gap: var(--base);
+    justify-content: space-between;
+}
+
 .form {
     display: flex;
     flex-direction: column;
+}
+
+.description {
+    margin: var(--base) 0;
 }
 
 .password {
@@ -46,47 +56,5 @@
 
     @include mid-break {
         margin: calc(var(--base) * 2) 0;
-    }
-}
-
-.buttonWrap {
-    display: flex;
-    gap: var(--base);
-    justify-content: space-between;
-}
-
-.warning, .emailInput {
-    margin-bottom: var(--base);
-}
-
-.modal {
-    & > * {
-        width: 800px;
-        overflow-x: scroll;
-        max-height: 100vh;
-    }
-}
-
-.modalActions {
-  display: flex;
-  gap: var(--base);
-  justify-content: flex-end;
-  width: 100%;
-  padding-top: calc(var(--base)*2);
-  border-top: 1px solid var(--theme-border-color);
-  margin-top: calc(var(--base)*2);
-
-}
-
-@include small-break {
-    .buttonWrap, .modalActions {
-        flex-wrap: wrap;
-    }
-
-    .modal {
-        & > * {
-            padding: calc(var(--base)*2) var(--gutter-h);
-            border: 0;
-        }
     }
 }
