+++ src/forms/Form/index.tsx
@@ -86,16 +86,18 @@ const Form = forwardRef<HTMLFormElement, FormProps>((props, ref) => {
             unflattenedData: reduceFieldsToValues(fields, true),
             dispatchFields: contextRef.current.dispatchFields,
           })
+
+          setHasSubmitted(false)
+          setIsModified(false)
+          setIsProcessing(false)
         } catch (err: unknown) {
           const message = err instanceof Error ? err.message : 'Unknown error'
           console.error(message) // eslint-disable-line no-console
+          setIsProcessing(false)
           setErrorFromSubmit(message)
         }
       }
 
-      setIsProcessing(false)
-      setIsModified(false)
-      setHasSubmitted(false)
       return false
     },
     [onSubmit, setHasSubmitted, setIsProcessing, setIsModified, fields],
